cyclocomp_package_dir {cyclocomp}	R Documentation

Cyclomatic complexity of a local package

Description

Automatically builds the package and installs it to a temporary directory.

Usage

cyclocomp_package_dir(path = ".")

Arguments

path
Path to the root directory of the R package.

Value

Data frame with two columns: name and cyclocomp.

See Also

Other cyclomatic complexity: cyclocomp_package(), cyclocomp()

[Package cyclocomp version 1.1.1 Index]